Refactoring related to all configuration files. Dropped the data variable. Moved all configuration interfaces to its own folder. Initiated work to reduce repeated code and started using base classes for Player and Enemies
