+++ ts/Functions.ts
@@ -1,18 +1,22 @@
 import { MultipackLoader } from "./Extensions/MultipackLoader";
 import { SpriteLoader } from "./Extensions/SpriteLoader";
 import { Game } from "./Game";
+import { AnimationSpriteConfig } from "./Interfaces/AnimationSpriteConfig";
+import { BackgroundConfig } from "./Interfaces/BackgroundsConfig";
+import { BackgroundSpritesConfig } from "./Interfaces/BackgroundSpritesConfig";
+import { CharacterAnimationDetailsConfig } from "./Interfaces/CharacterAnimationDetailsConfig";
+import { CharacterAnimationStatesConfig } from "./Interfaces/CharacterAnimationStatesConfig";
+import { CharacterConfig } from "./Interfaces/CharactersConfig";
+import { EntityConfig } from "./Interfaces/EntitiesConfig";
+import { LevelCharacterConfig } from "./Interfaces/LevelCharacterConfig";
+import { LevelConfig } from "./Interfaces/LevelsConfig";
+import { SoundConfig } from "./Interfaces/SoundsConfig";
 import { AnimationDetails, AnimationSprite, AnimationSprites, AnimationState } from "./Models/AnimatedSprite";
 import { Background } from "./Models/Background";
 import { Character, CharacterAction } from "./Models/Character";
-import { AnimationSpriteConfig, AnimationSpritesConfig } from "./Models/Configuration/AnimationSpritesConfig";
-import { BackgroundConfig, BackgroundsConfig, BackgroundSpritesConfig } from "./Models/Configuration/BackgroundsConfig";
-import { CharacterAnimationDetailsConfig, CharacterConfig, CharactersConfig, CharacterAnimationStatesConfig } from "./Models/Configuration/CharactersConfig";
-import { EntitiesConfig, EntityConfig } from "./Models/Configuration/EntitiesConfig";
-import { LevelCharacterConfig, LevelConfig, LevelsConfig } from "./Models/Configuration/LevelsConfig";
-import { EntitySound, Entity } from "./Models/Entities";
+import { Entity, EntitySound } from "./Models/Entities";
 import { LevelData } from "./Models/Level";
-import { Sounds, Sound } from "./Models/Sound";
-import { SoundsConfig, SoundConfig } from "./Models/Configuration/SoundsConfig";
+import { Sound, Sounds } from "./Models/Sound";
 
 /**
  * Loads a local .json file and returns the contents of the file
@@ -42,18 +46,18 @@ export function loadSounds(app: PIXI.Application): Promise<Sounds> {
 
   return new Promise(async (resolve) => {
     let sounds: Sounds = new Sounds();
-    let soundDataData: SoundsConfig;
+    let soundDataData: SoundConfig[];
 
     loadJSON('config/sounds.json')
       .then(data => {
-        soundDataData = JSON.parse(data) as SoundsConfig;
+        soundDataData = JSON.parse(data) as SoundConfig[];
 
         if (soundDataData == null) {
           throw new Error('Unable to load sounds');
         }
 
-        for (let i = 0; i < soundDataData.data.length; i++) {
-          let config: SoundConfig = soundDataData.data[i];
+        for (let i = 0; i < soundDataData.length; i++) {
+          let config: SoundConfig = soundDataData[i];
 
           // Prepare sound effect
           let sound: Sound = new Sound(config.id, config.filename);
@@ -73,18 +77,18 @@ export function loadBackgrounds(app: PIXI.Application): Promise<Map<string, Back
 
   return new Promise(async (resolve) => {
     let backgrounds: Map<string, Background> = new Map<string, Background>();
-    let backgroundData: BackgroundsConfig;
+    let backgroundData: BackgroundConfig[];
 
     loadJSON('config/backgrounds.json')
       .then(data => {
-        backgroundData = JSON.parse(data) as BackgroundsConfig;
+        backgroundData = JSON.parse(data) as BackgroundConfig[];
 
         if (backgroundData == null) {
           throw new Error('Unable to load backgrounds');
         }
 
-        for (let i = 0; i < backgroundData.data.length; i++) {
-          let config: BackgroundConfig = backgroundData.data[i];
+        for (let i = 0; i < backgroundData.length; i++) {
+          let config: BackgroundConfig = backgroundData[i];
 
           // Prepare background
           let background: Background = new Background(config.name);
@@ -122,19 +126,19 @@ export function loadBackgrounds(app: PIXI.Application): Promise<Map<string, Back
 export async function loadAnimationSprites(): Promise<AnimationSprites> {
 
   return new Promise(async (resolve) => {
-    let animationData: AnimationSpritesConfig;
+    let animationData: AnimationSpriteConfig[];
     let animationSprites: AnimationSprites = new AnimationSprites();
 
     loadJSON('config/animation_sprites.json')
       .then(data => {
-        animationData = JSON.parse(data) as AnimationSpritesConfig;
+        animationData = JSON.parse(data) as AnimationSpriteConfig[];
 
         if (animationData == null) {
           throw new Error('Unable to load animation data.');
         }
 
-        for (let i = 0; i < animationData.data.length; i++) {
-          let config: AnimationSpriteConfig = animationData.data[i];
+        for (let i = 0; i < animationData.length; i++) {
+          let config: AnimationSpriteConfig = animationData[i];
 
           // Diagnostics
           console.info(`Loading ${config.id} ${config.filename}`);
@@ -144,7 +148,7 @@ export async function loadAnimationSprites(): Promise<AnimationSprites> {
           packLoader.loadFile(config.filename, config.startAt, config.endAt, (id: string, message: string, sprites: AnimationSprite) => {
             animationSprites.data.set(config.id, sprites);
 
-            if (animationData.data.length == animationSprites.data.size) {
+            if (animationData.length == animationSprites.data.size) {
               resolve(animationSprites);
             }
           });
@@ -157,11 +161,11 @@ export async function loadAnimationSprites(): Promise<AnimationSprites> {
 export async function loadEntities(): Promise<Map<string, Entity>> {
 
   return new Promise(async (resolve) => {
-    let entityData: EntitiesConfig;
+    let entityData: EntityConfig[];
 
     loadJSON('config/entities.json')
       .then(data => {
-        entityData = JSON.parse(data) as EntitiesConfig;
+        entityData = JSON.parse(data) as EntityConfig[];
 
         if (entityData == null) {
           throw new Error('Unable to load entities data.');
@@ -171,8 +175,8 @@ export async function loadEntities(): Promise<Map<string, Entity>> {
         let loader: SpriteLoader = new SpriteLoader();
 
         // Get all config items
-        for (let i = 0; i < entityData.data.length; i++) {
-          let config: EntityConfig = entityData.data[i];
+        for (let i = 0; i < entityData.length; i++) {
+          let config: EntityConfig = entityData[i];
 
           // Diagnostics
           console.info(`Adding ${config.id} ${config.filename}`);
@@ -199,18 +203,18 @@ export function loadCharacters(game: Game): Promise<Map<string, Character>> {
   return new Promise(async (resolve) => {
 
     let characters: Map<string, Character> = new Map<string, Character>();
-    let characterData: CharactersConfig;
+    let characterData: CharacterConfig[];
 
     loadJSON('config/characters.json')
       .then(data => {
-        characterData = JSON.parse(data) as CharactersConfig;
+        characterData = JSON.parse(data) as CharacterConfig[];
 
         if (characterData == null) {
           throw new Error('Unable to load character data');
         }
 
-        for (let i = 0; i < characterData.data.length; i++) {
-          let config: CharacterConfig = characterData.data[i];
+        for (let i = 0; i < characterData.length; i++) {
+          let config: CharacterConfig = characterData[i];
           let animationSource: AnimationSprite;
 
           // Set animation source
@@ -288,18 +292,18 @@ export function loadLevels(app: PIXI.Application, game: Game): Promise<LevelData
 
   return new Promise(async (resolve) => {
     let levels: LevelData[] = [];
-    let levelData: LevelsConfig;
+    let levelData: LevelConfig[];
 
     loadJSON('config/levels.json')
       .then(data => {
-        levelData = JSON.parse(data) as LevelsConfig;
+        levelData = JSON.parse(data) as LevelConfig[];
 
         if (levelData == null) {
           throw new Error('Unable to load levels');
         }
 
-        for (let i = 0; i < levelData.data.length; i++) {
-          let config: LevelConfig = levelData.data[i];
+        for (let i = 0; i < levelData.length; i++) {
+          let config: LevelConfig = levelData[i];
 
           // Create level object
