+++ ts/Interfaces/BackgroundSpritesConfig.ts
@@ -1,14 +1,4 @@
 import { Point } from "pixi.js";
-
-export interface BackgroundsConfig {
-    data: BackgroundConfig[];
-}
-
-export interface BackgroundConfig {
-    name: string;
-    sprites: BackgroundSpritesConfig[];
-}
-
 export interface BackgroundSpritesConfig {
     file: string;
     index: number;
