+++ ts/Interfaces/BackgroundsConfig.ts
@@ -1,0 +1,7 @@
+import { BackgroundSpritesConfig } from "./BackgroundSpritesConfig";
+
+export interface BackgroundConfig {
+    name: string;
+    sprites: BackgroundSpritesConfig[];
+}
+
