+++ ts/Interfaces/CharacterAnimationDetailsConfig.ts
@@ -1,0 +1,5 @@
+import { CharacterAnimationDetailsDataConfig } from "./CharacterAnimationDetailsDataConfig";
+export interface CharacterAnimationDetailsConfig {
+    key: string;
+    overrides: CharacterAnimationDetailsDataConfig;
+}
