+++ ts/Interfaces/CharacterAnimationDetailsDataConfig.ts
@@ -1,0 +1,4 @@
+export interface CharacterAnimationDetailsDataConfig {
+    animationSpeed: number;
+    loop: boolean;
+}
