+++ ts/Interfaces/CharacterAnimationStatesConfig.ts
@@ -1,0 +1,4 @@
+export interface CharacterAnimationStatesConfig {
+    state: string;
+    animationKey: string;
+}
