+++ ts/Interfaces/EntitySoundConfig.ts
@@ -1,0 +1,4 @@
+export interface EntitySoundConfig {
+    id: string;
+    volume: number;
+}
