+++ ts/Interfaces/LevelCharacterShieldConfig.ts
@@ -1,0 +1,4 @@
+export interface LevelCharacterShieldConfig {
+    strength: number;
+    rechargeRate: number;
+}
