+++ ts/Interfaces/LevelsConfig.ts
@@ -1,0 +1,9 @@
+import { LevelCharacterConfig } from "./LevelCharacterConfig";
+
+export interface LevelConfig {
+    name: string;
+    background: string;
+    backgroundMusic: string;
+    characters: LevelCharacterConfig[];
+}
+
