+++ ts/Interfaces/SoundsConfig.ts
@@ -1,7 +1,3 @@
-export interface SoundsConfig {
-    data: SoundConfig[];
-}
-
 export interface SoundConfig {
     id: string;
     filename: string;
